DFQ/requirements.txt
pydot==1.4.1
torch==1.1.0
matplotlib==3.1.0
scipy==1.3.0
numpy==1.16.4
torchvision==0.3.0
graphviz==0.10.1
Pillow==6.2.1
